BAP-21019: Switch to PSR-12 coding standard except explicit constants visibility requirement (#35883)
- used PSR-12 except required visibility for constants
- updated php-cs-fixer config
- added doc block comments for changed classes
- updated php-semver-checker

+++ src/Oro/Bundle/OrderBundle/Form/Type/OrderDiscountItemType.php
@@ -173,7 +173,7 @@ class OrderDiscountItemType extends AbstractType
         $form = $event->getForm();
         $data = $event->getData();
         if ($data && $form->has('value') && null !== $data->getValue()) {
-            $form->get('value')->setData((double)$data->getValue());
+            $form->get('value')->setData((float)$data->getValue());
         }
     }
